WIP move net.imglib2.ui classes
PainterThread -> bdv.viewer.render
