+++ src/main/java/bdv/viewer/ViewerPanel.java
@@ -82,7 +82,7 @@ import net.imglib2.RealLocalizable;
 import net.imglib2.RealPoint;
 import net.imglib2.RealPositionable;
 import net.imglib2.realtransform.AffineTransform3D;
-import net.imglib2.ui.PainterThread;
+import bdv.viewer.render.PainterThread;
 import net.imglib2.util.LinAlgHelpers;
 import org.scijava.listeners.Listeners;
